feature: moved functions to modules
Refs #2
